Ongoing refactor, still working
