+++ typescript/tslint.json
@@ -5,7 +5,7 @@
     "align": [true, "parameters"],
     "class-name": true,
     "indent": [true, "spaces"],
-    "max-line-length": [true, 100],
+    "max-line-length": [true, 150],
     "no-consecutive-blank-lines": true,
     "no-trailing-whitespace": true,
     "no-duplicate-key": true,
